+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.12.0
+
+* Adds support for marker clustering.
+
 ## 2.11.0
 
 * Adds support for heatmap layers.